+++ aquadoggo/src/graphql/types/mod.rs
@@ -1,7 +1,11 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+mod document;
+mod document_fields;
 mod document_meta;
 mod next_arguments;
 
+pub use document::Document;
+pub use document_fields::DocumentFields;
 pub use document_meta::DocumentMeta;
 pub use next_arguments::NextArguments;