+++ aquadoggo/src/graphql/utils.rs
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-use async_graphql::dynamic::ResolverContext;
+use async_graphql::dynamic::{ResolverContext, TypeRef};
 use async_graphql::Value;
 use p2panda_rs::document::{DocumentId, DocumentViewId};
 use p2panda_rs::operation::OperationValue;
-use p2panda_rs::schema::SchemaId;
+use p2panda_rs::schema::{FieldType, SchemaId};
 use p2panda_rs::storage_provider::error::DocumentStorageError;
 use p2panda_rs::storage_provider::traits::DocumentStore;
 
@@ -31,9 +31,33 @@ pub fn gql_scalar(operation_value: &OperationValue) -> Value {
     }
 }
 
+/// Get the GraphQL type name for a p2panda field type.
+///
+/// GraphQL types for relations use the p2panda schema id as their name.
+pub fn graphql_type(field_type: &FieldType) -> TypeRef {
+    match field_type {
+        p2panda_rs::schema::FieldType::Boolean => TypeRef::named(TypeRef::BOOLEAN),
+        p2panda_rs::schema::FieldType::Integer => TypeRef::named(TypeRef::INT),
+        p2panda_rs::schema::FieldType::Float => TypeRef::named(TypeRef::FLOAT),
+        p2panda_rs::schema::FieldType::String => TypeRef::named(TypeRef::STRING),
+        p2panda_rs::schema::FieldType::Relation(schema_id) => TypeRef::named(schema_id.to_string()),
+        p2panda_rs::schema::FieldType::RelationList(schema_id) => {
+            TypeRef::named_list(schema_id.to_string())
+        }
+        p2panda_rs::schema::FieldType::PinnedRelation(schema_id) => {
+            TypeRef::named(schema_id.to_string())
+        }
+        p2panda_rs::schema::FieldType::PinnedRelationList(schema_id) => {
+            TypeRef::named_list(schema_id.to_string())
+        }
+    }
+}
+
 /// Downcast document id and document view id from parameters passed up the query fields and
 /// retrieved via the `ResolverContext`.
-pub fn downcast_id_params(
+///
+/// We unwrap internally here as we expect validation to have occured in the query resolver.
+pub fn downcast_document_id_arguments(
     ctx: &ResolverContext,
 ) -> (Option<DocumentIdScalar>, Option<DocumentViewIdScalar>) {
     ctx.parent_value
@@ -42,7 +66,7 @@ pub fn downcast_id_params(
         .to_owned()
 }
 
-/// Helper for getting a document from score by either the document id or document view id.
+/// Helper for getting a document from the store by either the document id or document view id.
 pub async fn get_document_from_params(
     store: &SqlStore,
     document_id: &Option<DocumentIdScalar>,